pyproject.toml
[tool.poetry]
name = "tzar"
version = "0.1.0"
description = ""
authors = ["Daniel Valenzuela <daniel@admetricks.com>"]

[tool.poetry.dependencies]
python = "^3.7"
click = "^8.0.1"
pyxdg = "^0.27"
toml = "^0.10.2"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[tool.poetry.scripts]
tzar = 'tzar.cli:run'

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
